// src/components/TypeDoc/api-docs.types.ts
export interface ApiDoc {
  classes?: ClassesItem[];
  enums?: EnumItem[];
  functions?: FunctionsItem[];
  interfaces?: InterfacesItem[];
  namespaces?: NamespaceItem[];
  typeAliases?: TypeAliasesItem[];
  variables?: VariableItem[];
}

export interface EnumItem {
  id: number;
  name: string;
}
export interface NamespaceItem {
  id: number;
  name: string;
}
export interface VariableItem {
  id: number;
  name: string;
}
export interface ClassesItem {
  id: number;
  name: string;
  source: Source;
  namespaceParentId: null;
  comment: Comment;
  external: boolean;
  'abstract': boolean;
  extendsType: null;
  implementsType: unknown[];
  typeParameters: TypeParametersItem[];
  construct: Construct;
  properties: PropertiesItem[];
  methods: MethodsItem[];
}
export interface Source {
  line: number;
  file: string;
  path: string;
  url: string;
}
export interface Comment {
  description: null;
  blockTags: unknown[];
  modifierTags: unknown[];
}
export interface TypeParametersItem {
  id: number;
  name: string;
  constraint: Constraint | null;
  'default': null;
}
export interface Constraint {
  kind: string;
  id: number;
  name: string;
  packageName: string;
  typeArguments: unknown[];
}
export interface Construct {
  id: number;
  name: string;
  source: Source | null;
  comment: Comment;
  parentId: number;
  accessibility: string;
  parameters: ParametersItem[];
}
export interface ParametersItem {
  id: number;
  name: string;
  comment: Comment;
  rest: boolean;
  optional: boolean;
  type: Type;
}
export interface Type {
  kind: string;
  type?: string | Type;
  id?: number | null;
  name?: string;
  packageName?: string | null;
  typeArguments?: TypeArgumentsItem[];
  properties?: null | PropertiesItem[];
  signatures?: null | SignaturesItem[];
  methods?: null;
  types?: TypesItem[];
}
export interface PropertiesItem {
  id: number;
  name: string;
  source?: Source;
  comment: Comment;
  parentId?: number;
  accessibility?: string;
  'abstract'?: boolean;
  'static'?: boolean;
  readonly: boolean;
  optional: boolean;
  type: Type;
}
export interface MethodsItem {
  id: number;
  name: string;
  source: Source;
  parentId: number;
  accessibility: string;
  'abstract': boolean;
  'static': boolean;
  signatures: SignaturesItem[];
}
export interface SignaturesItem {
  id: number;
  name: string;
  comment: Comment;
  typeParameters: TypeParametersItem[];
  parameters: ParametersItem[];
  returnType: ReturnType;
}
// ...
export interface FunctionsItem {
  id: number;
  name: string;
  source: Source;
  namespaceParentId: null;
  comment: Comment;
  external: boolean;
  signatures: SignaturesItem[];
}
export interface InterfacesItem {
  id: number;
  name: string;
  source: Source;
  namespaceParentId: null;
  comment: Comment;
  external: boolean;
  typeParameters: TypeParametersItem[];
  properties: PropertiesItem[];
  methods: unknown[];
}
export interface TypeAliasesItem {
  id: number;
  name: string;
  source: Source;
  namespaceParentId: null;
  comment: Comment;
  external: boolean;
  typeParameters: TypeParametersItem[];
  type: Type;
}

export type ItemsTypes = ClassesItem & EnumItem & FunctionsItem & InterfacesItem & NamespaceItem & TypeAliasesItem & VariableItem;